LayoutTests/http/tests/cookies/cookie-with-multiple-level-path-expected.txt

Based on RFC 6265 Section-5.4: 2. The user agent SHOULD sort the cookie-list in the following order: a) Cookies with longer paths are listed before cookies with shorter paths. b) Among cookies that have equal-length path fields, cookies with earlier creation-times are listed before cookies with later creation-times.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS () => result is 'foo=buz; foo=bar'
PASS successfullyParsed is true

TEST COMPLETE
